feat: separate clipboards for different views
