+++ packages/blocks/src/database-block/kanban/clipboard.ts
@@ -1,0 +1,7 @@
+import type { BaseViewClipboard } from '../common/clipboard.js';
+
+export class KanbanViewClipboard implements BaseViewClipboard {
+  init() {
+    // todo
+  }
+}
